Improve logic for detecting PKCS 12 keys
When attempting to load a PKCS 12 key in Python 3, io.open would
throw an error since it tried to apply UTF-8 text encoding to the
key file, which was not a compatible text file.

Additional functions were introduced to check if a file may be in
PKCS 12 format, and if so, use open(filename, 'rb') instead.